fixing wrong cast

+++ test/test_gpu2.cpp
@@ -9719,9 +9719,9 @@ TEST_F(NVFuserTest, FusionPersistentBufferProjection2_CUDA) {
   // projectable buffer inputs. Thus, the buffer size would be
   // calculated as the sum of tv1, tv0 and tv1.
   auto projected_size = persistent_buffer_size.projected_persistent_buffer_size;
-  auto expected_size = shape[1] * 2 * dataTypeSize(DataType::Half);
+  auto expected_size = static_cast<int64_t>(shape[1] * 2 * dataTypeSize(DataType::Half));
   TORCH_CHECK(
-      (int64_t)projected_size == expected_size,
+      projected_size == expected_size,
       "Buffer projection failure. Expected size: ",
       expected_size,
       ". Actual: ",
